+++ src/Plugins/Connect/MSFSSimConnectPlugin/src/Event/SimConnectEvent.h
@@ -93,8 +93,10 @@ namespace SimConnectEvent
         CustomReplay,
         CustomPause,
         CustomStop,
-        CustomForward,
-        CustomBackward,
+        CustomForwardDown,
+        CustomForwardUp,
+        CustomBackwardDown,
+        CustomBackwardUp,
         CustomBegin,
         CustomEnd
     };